% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{concentration}
\alias{concentration}
\title{concentration}
\usage{
concentration(level = 1)
}
\arguments{
\item{level}{An integer (1, 2, or 3) denoting the difficutly of the board. Defaults to 1 (easiest).}
}
\description{
Click on pairs of cards until all matches are found.
}
\examples{
concentration(2)
}
\keyword{concentration}
