% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathsquare.R
\name{mathsquare}
\alias{mathsquare}
\title{mathsquare}
\usage{
mathsquare()
}
\description{
Drag numbers into the grid until all rows, columns, and diagonals sum to the same value.  To move a value from the grid back to the starting area, drag a new number to the cell containing the number you wish to remove.  The undesirable number will be replaced and sent back to the starting area.
}
\keyword{mathsquare}
