% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathtriangle.R
\name{mathtriangle}
\alias{mathtriangle}
\title{mathtriangle}
\usage{
mathtriangle()
}
\description{
Drag numbers into the empty boxes until all sides of the triangle sum to the same value.  To move a value from the triangle back to the starting area, drag a new number to the cell containing the number you wish to remove.  The undesirable number will be replaced and sent back to the starting area.
}
\keyword{mathtriangle}
